Libft

+++ Libft/ft_bzero.c
@@ -1,6 +1,8 @@
-void	ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
-	unsigned int i;
+	size_t i;
 	unsigned char *str = (unsigned char *)s;
 
 	i = -1;
