+++ Libft/ft_memcmp.c
@@ -1,6 +1,8 @@
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 	unsigned char *str1 = (unsigned char *)s1;
 	unsigned char *str2 = (unsigned char *)s2;
