+++ Libft/ft_memmove.c
@@ -1,6 +1,8 @@
-void *ft_memmove(void *dst, const void *src, unsigned int len)
+#include "libft.h"
+
+void *ft_memmove(void *dst, const void *src, size_t len)
 {
-	unsigned int i;
+	size_t i;
 	unsigned char *d = (unsigned char *)dst;
 	unsigned char *s = (const unsigned char *)src;
