+++ Libft/ft_strlen.c
@@ -1,6 +1,8 @@
-unsigned int ft_strlen(const char *s)
+#include "libft.h"
+
+size_t ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while(s[i])
