+++ Libft/ft_strnstr.c
@@ -1,7 +1,9 @@
-char *ft_strnstr(const char *haystack, const char *needle, unsigned int len)
+#include "libft.h"
+
+char *ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned int i;
-	unsigned int s;
+	size_t i;
+	size_t s;
 
 	i = 0;
 	if (needle[0] == 0)
